[Feature] - implement the dump tasks for creditcard tables in TD

+++ app/dump.py
@@ -15,19 +15,64 @@ BASEPATH = "{}/..".format(os.path.dirname(os.path.abspath(__file__)))
 BASEPATH_TERADATA = os.path.join(BASEPATH, "data", "teradata")
 
 
+class DumpAllTask(luigi.Task):
+    task_namespace = "dump"
+
+    remove = luigi.BoolParameter()
+    today = luigi.DateParameter(default=datetime.datetime.now())
+
+    def requires(self):
+        yield DayNADumpTask(remove=self.remove)
+        yield MonthMonthDumpTask(remove=self.remove, today=self.today)
+        yield YearYearDumpTask(remove=self.remove, today=self.today)
+        yield MonthYearDumpTask(remove=self.remove, today=self.today)
+
+        yield DayMonthDumpTask(remove=self.remove, today=self.today.strftime("%Y-%m-%d"))
+
+class DayNADumpTask(luigi.Task):
+    task_namespace = "dump"
+
+    remove = luigi.BoolParameter()
+
+    def requires(self):
+        global BASEPATH_TERADATA
+
+        tables = ["VP_MCIF.PARTY_CC",
+                  "VP_MCIF.ACCT_INS_CATHAYLIFE",
+                  "VP_MCIF.ACCT_DRV_MF",
+                  "VP_MCIF.ACCT_INS_PROPERTY",
+                  "VP_MCIF.LOCATION_DRV",
+                  "VP_MCIF.RD_MIS_MCC_CODE",
+                  "VP_MCIF.RD_MIS_MCC_GROUP_CODE",
+                  "VP_MCIF.RD_MIS_MERCHAINT_ID",
+                  "VP_MCIF.RD_CC_CATHAY_CARD_TYPE",
+                  "VP_MCIF.EVENT_CC_AIRPORT_PARKING",
+                  "DP_MCIF_REF.RD_ABT_PROD_CODES"]
+
+        for table in tables:
+            sql = "SELECT * FROM {}".format(table)
+            ofile = os.path.join(BASEPATH_TERADATA, "{}.tsv.gz".format(table))
+
+            if self.remove and os.path.exists(ofile):
+                os.remove(ofile)
+
+            yield TeradataTable(query=sql, ofile=ofile)
+
 class DayMonthDumpTask(luigi.Task):
     task_namespace = "dump"
 
-    interval = luigi.DateIntervalParameter()
+    remove = luigi.BoolParameter()
+    today = luigi.Parameter()
 
     def requires(self):
         global BASEPATH_TERADATA
 
         sqls = ["SELECT * FROM VP_MCIF.PARTY_DRV",
-                "SELECT * FROM VP_MCIF.EVENT_CC_TXN"]
+                "SELECT CUSTOMER_ID, CARD_TYPE_CATEGORY_CODE,CARD_TYPE_CODE,CARD_NBR,TXN_DATE,TXN_CODE,TXN_AMT,MERCHANT_NBR,MERCHANT_CATEGORY_CODE,MERCHANT_NAME,MERCHANT_LOCATION_CITY,MERCHANT_LOCATION_COUNTRY_CODE,ORIGINAL_CURRENCY_CODE,TXN_AMT_US_DOLLAR,PRIMARY_CARDHOLDER_IND,TREATY_CONV_AMT FROM VP_MCIF.EVENT_CC_TXN"]
 
         done = set()
-        for month in self.interval:
+        for diff in [0, 30, 60, 90, 120, 150, 180]:
+            month = datetime.datetime.strptime(self.today, "%Y-%m-%d") - datetime.timedelta(days=diff)
             past = month.strftime("%Y%m")
 
             for sql in sqls[:]:
@@ -44,7 +89,7 @@ class DayMonthDumpTask(luigi.Task):
                         if month.day == 1:
                             ofile = os.path.join(BASEPATH_TERADATA, "{}.tsv.gz".format(table))
 
-                            if os.path.exists(ofile):
+                            if self.remove and os.path.exists(ofile):
                                 os.remove(ofile)
 
                             yield TeradataTable(query=sql, ofile=ofile)
@@ -78,17 +123,18 @@ class YearYearDumpTask(luigi.Task):
     task_namespace = "dump"
 
     remove = luigi.BoolParameter()
-    year = luigi.IntParameter(default=datetime.datetime.now().year-1)
+    today = luigi.DateParameter(default=datetime.datetime.now())
 
     def requires(self):
         global BASEPATH_TERADATA
 
+        year = self.today.year-1
         sqls = ["SELECT * FROM VP_MCIF.RPT_CC_COST", "SELECT * FROM VP_MCIF.RPT_CC_COST2"]
 
         for sql in sqls[:]:
             table = sql.split(" ")[-1]
             ofile = os.path.join(BASEPATH_TERADATA, "{}.tsv.gz".format(table))
-            sql = sql + str(self.year)
+            sql = sql + str(year)
 
             if self.remove and os.path.exists(ofile):
                 os.remove(ofile)
@@ -106,6 +152,7 @@ class MonthYearDumpTask(luigi.Task):
 
         sqls = ["SELECT * FROM VP_MCIF.PARTY_CC_JCIC_KRM046_DATA"]
 
+        self.today = self.today.replace(day=1)
         if self.today.day == 1:
             for sql in sqls[:]:
                 table = sql.split(" ")[-1]
@@ -118,5 +165,3 @@ class MonthYearDumpTask(luigi.Task):
                 yield TeradataTable(query=sql, ofile=ofile)
         else:
             logger.warn("not the first date of this month({})".format(self.today))
-
-
